Deleted ocs_ci/ocs/overprovision.py file and moved functions to the ocp.py file.
